+++ src/directives.js
@@ -11,9 +11,9 @@ angular.module('oi.select')
         scope: {},
         compile: function (element, attrs) {
             var optionsExp = attrs.oiOptions,
-                match;
+                match = optionsExp ? optionsExp.match(NG_OPTIONS_REGEXP) : ['', 'i', '', '', '', 'i', '', '', ''];
 
-            if (!(match = optionsExp.match(NG_OPTIONS_REGEXP))) {
+            if (!match) {
                 throw new Error("Expected expression in form of '_select_ (as _label_)? for (_key_,)?_value_ in _collection_'");
             }
 
@@ -514,7 +514,7 @@ angular.module('oi.select')
                 }
 
                 function getMatches(query, selectedAs) {
-                    var values = valuesFn(scope.$parent, {$query: query, $selectedAs: selectedAs}),
+                    var values = valuesFn(scope.$parent, {$query: query, $selectedAs: selectedAs}) || '',
                         waitTime = 0;
 
                     scope.selectorPosition = options.newItem === 'prompt' ? false : 0;
@@ -533,6 +533,12 @@ angular.module('oi.select')
 
                         return $q.when(values.$promise || values)
                             .then(function(values) {
+                                if (!values) {
+                                    scope.groups = [];
+                                    updateGroupPos();
+                                    return;
+                                }
+
                                 if (!selectedAs) {
                                     var outputValues = multiple ? scope.output : [];
                                     var filteredList   = listFilter(oiUtils.objToArr(values), query, getLabel, listFilterOptionsFn(scope.$parent));
@@ -540,7 +546,6 @@ angular.module('oi.select')
                                     var filteredOutput = filter(withoutOverlap);
 
                                     scope.groups = group(filteredOutput);
-
                                     updateGroupPos();
                                 }
                                 return values;