+++ CHANGELOG.md
@@ -2,7 +2,6 @@
 
 #### Features
 
-
 - **oi-select:** is optional (useful for new item case)
 - **multiple:** is observable
 - **multiple-limit:** is observable
@@ -30,6 +29,7 @@
     - cleanModel: changed behavior (clean model by click). It affected `newItem: 'prompt'`
 
 #### Bug Fixes
+
 - **dropdown** fixed dropdown opening in IE, open dropdown by pushing on letter keys (for single select)
 - **oi-select-options:**
     - closeList: fixed